requirements.txt
asgiref==3.5.2
dj-rest-auth==2.2.5
Django==4.1.2
django-extensions==3.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.4
pytz==2022.4
sqlparse==0.4.3
